services/ai/run.py
import uvicorn
import os
import dotenv

# Load environment variables
dotenv.load_dotenv()

if __name__ == "__main__":
    # Get configuration from environment variables
    host = os.getenv("HOST", "0.0.0.0")
    port = int(os.getenv("PORT", "8000"))
    reload = os.getenv("DEBUG", "false").lower() == "true"

    # Run the FastAPI app
    uvicorn.run(
        "src.main:app",
        host=host,
        port=port,
        reload=reload,
        log_level="info" if reload else "warning",
    )
