services/ai/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sentence-transformers==2.6.0
python-dotenv==1.0.1
pydantic>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.23.2 